services/rasa/requirements.txt
rasa-pro
rasa-sdk
openai
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
spacy==3.6.1
fastapi
uvicorn[standard]
httpx